CARD_extraction

Extracts variables from time series (for example, the yearly mean of a time series) using CARD parameterization files.

Usage

CARD_extraction(
  data,
  CARD_path = system.file("CARD", package = "EXstat"),
  CARD_tmp = NULL,
  CARD_dir = "__all__",
  CARD_name = NULL,
  period_default = NULL,
  suffix = NULL,
  suffix_delimiter = "_",
  cancel_lim = FALSE,
  simplify = FALSE,
  expand_overwrite = NULL,
  sampling_period_overwrite = NULL,
  rmNApct = TRUE,
  rm_duplicates = FALSE,
  dev = FALSE,
  verbose = FALSE
)

Arguments

data

Input data format is a tibble from the tibble package. It needs to have :

Only one column of Date that are regularly spaced and unique for each time serie.

If there is more than one time serie, at least one column needs to be of character string for names of time series in order to identify them. If more than one column of identifier is given, they will all be used in order to identify a unique time serie.

At least one column of numeric (or logical) on which the process of variable extraction will be perform. More numerical column can be leave but if they are useless, they will be suppressed.

e.g.

> data
A tibble: 201 × 4
   time         Q_obs  Q_sim  ID
   <date>       <dbl>  <dbl>  <chr>
1   2000-02-10   10     97.8  serie 1
2   2000-02-11   19    -20.5  serie 1
3   2000-02-12   13    -76.9  serie 1
4   2000-02-13   15    -86.0  serie 1
    ...
103 2001-01-01  1.3     1988  serie 2
104 2001-01-02  1.2      109  serie 2
105 2001-01-03  1.0       90  serie 2
106 2001-01-04  1.1       91  serie 2
    ...

CARD_path

A character string representing the path to the downloaded CARD directory (it should end with "CARD"). In this directory, you can copy and paste (and later modify) CARDs from the "__all__" subdirectory that you want to use for an analysis represented by a subdirectory named CARD_dir (see CARD_tmp if you want to locate your CARD_dir directory elsewhere). In your CARDs, you can specify functions available in the scripts of the "__tools__" subdirectory.

CARD_tmp

If you want to locate the CARD_dir directory somewhere other than in the CARD_path directory, you can specify a character string in CARD_tmp for a path where the CARD_dir subdirectory of CARDs will be searched. Default is "NULL" if you want to locate the CARD_dir subdirectory of CARDs in CARD_path.

CARD_dir

A character string for the name of a subdirectory in CARD_path (or CARD_tmp) where the CARD parameterization files are located for an analysis. Default is "WIP".

CARD_name

By default, all CARDs in the CARD_dir directory will be used for the analysis. However, you can specify a vector of character strings with the names of the CARDs to be used. Default is "NULL" for using all the CARDs.

period_default

A vector of two dates (or two unambiguous character strings that can be coerced to dates) to restrict the period of analysis. As an example, it can be c("1950-01-01", "2020-12-31") to select data from the 1st January of 1950 to the end of December of 2020. The default option is period=NULL, which considers all available data for each time serie.

suffix

A character string vector representing suffixes to be appended to the column names of the extracted variables. This parameter allows handling multiple extraction scenarios. For example, a cumbersome case can be to have a unique function to apply to a multiple list of column. It is possible to give funct=list(QA_obs=mean, QA_sim=mean) and funct_args=list(list("Q_obs", na.rm=TRUE), list("Q_sim", na.rm=TRUE)) or simply funct=list(QA=mean) and funct_args=list("Q", na.rm=TRUE) with suffix=c("obs", "sim"). The two approach give the same result. Default NULL.

suffix_delimiter

character string specifies the delimiter to use between the variable name and the suffix if not NULL. The default is "_".

cancel_lim

A logical to specify whether to cancel the NA percentage limits in the CARDs. Default is FALSE.

simplify

A logical to specify whether to simplify the extracted data by joining each tibble extracted from each CARDs. Usefull when the extracted variable has no temporal extension. Default "FALSE".

expand_overwrite

logical or NULL. If TRUE, expand the output tibble as a list of tibble for each extracted variable by suffix. Default NULL to conserve the value specified in the CARDs used.

sampling_period_overwrite

A character string or a vector of two character strings that will indicate how to sample the data for each time step defined by time_step. Hence, the choice of this argument needs to be link with the choice of the time step. For example, for a yearly extraction so if time_step is set to "year", sampling_period needs to be formated as %m-%d (a month - a day of the year) in order to indicate the start of the sampling of data for the current year. More precisly, if time_step="year" and sampling_period="03-19", funct will be apply on every data from the 3rd march of each year to the 2nd march of the following one. In this way, it is possible to create a sub-year sampling with a vector of two character strings as sampling_period=c("02-01", "07-31") in order to process data only if the date is between the 1st february and the 31th jully of each year. not available for now For a monthly (or seasonal) extraction, sampling_period needs to give only day in each month, so for example sampling_period="10" to extract data from the 10th of each month to the 9th of each following month. Default NULL to conserve the value specified in the CARDs used.

rmNApct

logical. Should the NApct column, which shows the percentage of missing values in the output, be removed ? Default TRUE.

rm_duplicates

logical. Should duplicate time series values be automatically removed ? Default FALSE.

dev

logical If TRUE, development mode is enabled. Default is FALSE.

verbose

logical. Should intermediate messages be printed during the execution of the function ? Default FALSE.

Value

A list of two tibbles.

The dataEX tibble, which contains the extracted variable, or a named list of tibbles for each extracted variable if expand_overwrite is TRUE.

The metaEX tibble, which contains the metadata of the extraction from CARDs.

See also

Examples

## Creation of random data set
set.seed(99)
Start = as.Date("2000-02-01")
End = as.Date("2010-04-04")
Date = seq.Date(Start, End, by="day")

# First time serie
data_1 = dplyr::tibble(time=Date,
                       X_state1=as.numeric(Date) +
                           rnorm(length(Date), 1e4, 1e3),
                       X_state2=seq(1, length(Date))/1e2 +
                           rnorm(length(Date), 0, 1),
                       id="serie 1")
data_1$X_state2[round(runif(500, 1, nrow(data_1)))] = NA

# Second time serie
data_2 = dplyr::tibble(time=Date,
                       X_state1=as.numeric(Date) +
                           rnorm(length(Date), 1e4, 1e3),
                       X_state2=seq(1, length(Date))/1e2 +
                           rnorm(length(Date), 0, 1),
                       id="serie 2")
data_2$X_state2[round(runif(1000, 1, nrow(data_2)))] = NA

# Final data for testing
data = dplyr::bind_rows(data_1, data_2)

## Extraction with CARD
# Copy and paste CARD from __all__ to the CARD_dir directory (or
# use CARD_tmp with CARD_management function) and then process the
# extraction
if (FALSE) { # \dontrun{
CARD_extraction(data,
                CARD_path="path/to/CARD",
                CARD_tmp="path/to/temporary",
                CARD_dir="WIP",
                period_default=c("1950-01-01", "2020-12-31"),
                simplify=FALSE,
                cancel_lim=TRUE,
                verbose=TRUE)
} # }