General Mann-Kendall

A general version of the Mann-Kendall test, enabling various dependence assumptions.

Usage

generalMannKendall(
  X,
  level = 0.1,
  dep.option = "INDE",
  DoDetrending = TRUE,
  verbose = FALSE
)

Arguments

X

numeric vector, data. IMPORTANT: X is assumed to be regularly-spaced. It uses NA to fill the gaps rather than removing missing values.

level

numeric in (0,1), level of the test.

dep.option

string, option for handling temporal dependence. Available :

'INDE', assume independence (i.e. the standard MK test)

'AR1', assumes AR1 short-term dependence structure (i.e. Hamed and Rao's version of the MK test)

'LTP', assume long-term persistence (i.e. Hamed's version of the MK test)

DoDetrending,

logical, only used for dep.option == LTP: do detrending before estimating Hurst coefficient (default=TRUE as recommended in Hamed's paper)

verbose

logical Whether to print intermediate messages. Default is FALSE.

Value

A list with the following fields :

H: logical, reject (true) or do not reject (false) H0

P: p-value of the test

STAT: test statistics

TREND: trend estimate (using Sen's slope estimate)

DEP: dependence estimate (= 0 if dep.option='INDE', =lag-1 autocorrelation if dep.option='AR1', =Hurst coefficient if dep.option='LTP')

Details

Handling of ties: Specific formula exist for INDE and AR1, but the LTP case is trickier. Hammed's paper is unclear on how to handle ties, especially at the step of Hurst coefficient estimation. There is a normal-score transformation at this step, and one needs to decide how to assign a rank to ties. What is implemented below is the option ties.method = "random", i.e. the rank is randomized for ties. This is not, strictly speaking, correct because this randomization impacts the dependence structure. However synthetic runs suggest it works OK.

Computational efficiency: Likely poor for case dep.option='LTP'. There is a 4-level loop which leads to a n^4 algorithm. I attempted to vectorize this loop but it didn't improve things => Expect significant running times for dep.option='LTP ' when size(X) > 50... (orders of magnitude: 1s for n=30, 10s for n=50, 2-3 minutes for n=100). On the other hand both options INDE and AR1 are very fast.

References

Hamed, Rao, 1998. A modified Mann-Kendall trend test for autocorrelated data. J. Hydrol., 204(1-4): 182-196.

Hamed, 2008. Trend detection in hydrologic data: The Mann-Kendall trend test under the scaling hypothesis. J. Hydrol., 349(3-4): 350-363.

Examples

if (FALSE) { # \dontrun{
data(nhtemp) #Average Yearly Temperatures in New Haven
generalMannKendall(X=nhtemp, dep.option='AR1')
} # }