process_extraction

Extracts a variable from time series (for example the yearly mean of time series). Extraction can have a specific time step and sampled differently along this time step.

Usage

process_extraction(
  data,
  funct = max,
  funct_args = list(),
  time_step = "year",
  sampling_period = NULL,
  period = NULL,
  is_date = FALSE,
  NApct_lim = NULL,
  NAyear_lim = NULL,
  Seasons = c("DJF", "MAM", "JJA", "SON"),
  nameEX = "X",
  suffix = NULL,
  suffix_delimiter = "_",
  keep = NULL,
  compress = FALSE,
  expand = FALSE,
  rmNApct = TRUE,
  rm_duplicates = FALSE,
  dev = FALSE,
  verbose = FALSE
)

Arguments

data

Input data format is a tibble from the tibble package. It needs to have :

Only one column of Date that are regularly spaced and unique for each time serie.

If there is more than one time serie, at least one column needs to be of character string for names of time series in order to identify them. If more than one column of identifier is given, they will all be used in order to identify a unique time serie.

At least one column of numeric (or logical) on which the process of variable extraction will be perform. More numerical column can be leave but if they are useless, they will be suppressed.

e.g.

> data
A tibble: 201 × 4
   time         Q_obs  Q_sim  ID
   <date>       <dbl>  <dbl>  <chr>
1   2000-02-10   10     97.8  serie 1
2   2000-02-11   19    -20.5  serie 1
3   2000-02-12   13    -76.9  serie 1
4   2000-02-13   15    -86.0  serie 1
    ...
103 2001-01-01  1.3     1988  serie 2
104 2001-01-02  1.2      109  serie 2
105 2001-01-03  1.0       90  serie 2
106 2001-01-04  1.1       91  serie 2
    ...

funct

The function that you want to use for the process of variable extraction. More specificaly, it is possible to give a list with several functions as element of that list and the name that will be used for the extracted column as the names element of each function of that previously defined list. A simple case will be funct=mean and a more complicated one funct=list(QA=mean, QJXA=max). Default max.

funct_args

A list of list of named arguments needed for each functions provided through funct. This list can be a simple list if there is only one function given by funct. The argument can relate to a column name in order to specify on which numerical column the extraction will be perfom. For the simple example, funct_args=list("Q_obs", na.rm=TRUE) and for the more complex case funct_args=list(list("Q_obs", na.rm=TRUE), list("Q_sim", na.rm=FALSE)). Default list.

time_step

A character string specifying the time step of the variable extraction process. Possible values are :

"year" for a value per year

"month" for a value for each month of the year (so 12 values if at least a full year is given)

"year-month" for a value for each month of each year (so 12 times the number of given year values at the end)

"season" for a value for each season of th year (so by default 4 values)

"year-season" for a value for each season of each year (so by default 4 times the number of given year values at the end)

"yearday" for one value per day of the year (so 365 values at the end if at least a full year is given... but more than one year seems obviously more interesting) "none" if you want to extract a unique value for the whole time serie Default "year".

sampling_period

A character string or a vector of two characters strings that will indicate how to sample the data for each time step defined by time_step. Hence, the choice of this argument needs to be link with the choice of the time step. For example, for a yearly extraction so if time_step is set to "year", sampling_period needs to be formated as %m-%d (a month - a day of the year) in order to indicate the start of the sampling of data for the current year. More precisly, if time_step="year" and sampling_period="03-19", funct will be apply on every data from the 3rd march of each year to the 2nd march of the following one. In this way, it is possible to create a sub-year sampling with a vector of two character strings as sampling_period=c("02-01", "07-31") in order to process data only if the date is between the 1st february and the 31th jully of each year. not available for now For a monthly (or seasonal) extraction, sampling_period needs to give only day in each month, so for example sampling_period="10" to extract data from the 10th of each month to the 9th of each following month. Default NULL.

period

A vector of two dates (or two unambiguous character strings that can be coerced to dates) to restrict the period of analysis. As an example, it can be c("1950-01-01", "2020-12-31") to select data from the 1st January of 1950 to the end of December of 2020. The default option is period=NULL, which considers all available data for each time serie.

is_date

logical. If TRUE, process_extration() will convert the result of the application of funct to a day of the year. The aim is for example to give funct=which.min and if is_date=TRUE, the result will not be the indice of the minimum of the sample but the associated day of the year given by an integer (1 is the 1st of january). Default FALSE.

NApct_lim

numeric. The maximum percentage of missing values in each sample allowed. If this threshold is exceeded, the value associated to the current sample will be convert to NA. Default NULL.

NAyear_lim

numeric.The maximum number of continuous missing years allowed. If this threshold is exceeded, the time serie will be split in half around the problematic period and only the longest part will be used for the extraction process. Default NULL.

Seasons

A vector of character strings that indicates the seasonal pattern of a year. All months of the year needs to be contain in the Seasons variable. Give months circulary in a vector in which each element is a character chain of several months identify by the first letter of their names. The default is Seasons=c("DJF", "MAM", "JJA", "SON") but it can be set for example to Seasons=c("MAMJJA", "SONDJF").

nameEX

A character string specifying the name of the column of the extracted variable if no name is given in funct. Default is "X".

suffix

A character string vector representing suffixes to be appended to the column names of the extracted variables. This parameter allows handling multiple extraction scenarios. For example, a cumbersome case can be to have a unique function to apply to a multiple list of column. It is possible to give funct=list(QA_obs=mean, QA_sim=mean) and funct_args=list(list("Q_obs", na.rm=TRUE), list("Q_sim", na.rm=TRUE)) or simply funct=list(QA=mean) and funct_args=list("Q", na.rm=TRUE) with suffix=c("obs", "sim"). The two approach give the same result. Default NULL.

suffix_delimiter

character string specifies the delimiter to use between the variable name and the suffix if not NULL. The default is "_".

keep

in developpement A character string vector of column names to keep in the output tibble. In the current state, keep can only be set to NULL if you don't want to keep anythings in the output besides the usefull column, or "all" if you want to conserve all the initial column in the output column. Warning : The number of rows in the output with keep="all" will, as a result, be the same as in the input. For example, the extracted value for a year from a daily time series will be assigned to the first day of that year, and NaN will be assigned to every other value in the output. Default NULL.

compress

logical. If time_step is set to "month", "year-month", "season" or "year-season" should the function return a standard tibble or a compressed one ? When compress = TRUE, the function will perform a pivot_wider operation to display the month or season information in columns instead of rows. Default FALSE.

e.g.

# for time_step="season" and compress=FALSE
# A tibble: 8 × 4
  ID       <time    Q    NApct
  <chr>    <chr>  <dbl>  <dbl>
1 serie 1  DJF    1464    8.6
2 serie 1  JJA    1447    0
3 serie 1  MAM    1395    0
4 serie 1  SON    1458    0
5 serie 2  DJF      11    8.6
6 serie 2  JJA       2    0
7 serie 2  MAM       1    0
8 serie 2  SON       4    0

# for time_step="season" and compress=TRUE
# A tibble: 2 × 5
 ID       Q_DJF  Q_JJA  Q_MAM  Q_SON
 <chr>    <dbl>  <dbl>  <dbl>  <dbl>
1 serie 1  1464  1447   1395   1458
2 serie 2    11     2      1      4

expand

logical. If TRUE, expand the output tibble as a list of tibble for each extracted variable by suffix. Default FALSE.

rmNApct

logical. Should the NApct column, which shows the percentage of missing values in the output, be removed ? Default TRUE.

rm_duplicates

logical. Should duplicate time series values be automatically removed ? Default FALSE.

dev

logical If TRUE, development mode is enabled. Default is FALSE.

verbose

logical. Should intermediate messages be printed during the execution of the function ? Default FALSE.

Value

A tibble containing the extracted variable, or a named list of tibble for each extracted variable if expand is TRUE. This output follows the same format as the input data described in data, making it possible to iterate over this output using process_extraction().

Note

compress and expand cannot be both set to TRUE for time_step set to "year-month" or "year-season".

NA values are considered missing values and are used to compute the percentage of gaps over each time step, potentially removing the results if the percentage exceeds NApct_lim.

NaN values are considered non-existent values but can also be computation artifacts, for example, when the keep option is used. For speed performance reasons, NaN values are needed as masked values for input time series that are not daily and are extracted throughout the year.

See also

Examples

## Creation of random data set
set.seed(99)
Start = as.Date("2000-02-01")
End = as.Date("2010-04-04")
Date = seq.Date(Start, End, by="day")

# First time serie
data_1 = dplyr::tibble(time=Date,
                       X_state1=as.numeric(Date) +
                           rnorm(length(Date), 1e4, 1e3),
                       X_state2=seq(1, length(Date))/1e2 +
                           rnorm(length(Date), 0, 1),
                       id="serie 1")
data_1$X_state2[round(runif(500, 1, nrow(data_1)))] = NA

# Second time serie
data_2 = dplyr::tibble(time=Date,
                       X_state1=as.numeric(Date) +
                           rnorm(length(Date), 1e4, 1e3),
                       X_state2=seq(1, length(Date))/1e2 +
                           rnorm(length(Date), 0, 1),
                       id="serie 2")
data_2$X_state2[round(runif(1000, 1, nrow(data_2)))] = NA

# Final data for testing
data = dplyr::bind_rows(data_1, data_2)

## Extraction of the yearly average of daily value.
process_extraction(data=data,
                   funct=list(XA_state1=mean),
                   funct_args=list("X_state1", na.rm=TRUE),
                   time_step="year")
#> # A tibble: 22 × 3
#>    id      time       XA_state1
#>    <chr>   <date>         <dbl>
#>  1 serie 1 2000-01-01    21092.
#>  2 serie 1 2001-01-01    21468.
#>  3 serie 1 2002-01-01    21895.
#>  4 serie 1 2003-01-01    22232.
#>  5 serie 1 2004-01-01    22656.
#>  6 serie 1 2005-01-01    22999.
#>  7 serie 1 2006-01-01    23357.
#>  8 serie 1 2007-01-01    23605.
#>  9 serie 1 2008-01-01    24223.
#> 10 serie 1 2009-01-01    24421.
#> # ℹ 12 more rows

## Extraction of the median of the yearly maximum of the monthly
# averages of daily values.
# Three steps are needed for this case, the first one is the
# monthly average over years.
dataEX_tmp =
    process_extraction(data=data,
                       funct=list(X_month_state2=mean),
                       funct_args=list("X_state2", na.rm=TRUE),
                       time_step="year-month",
                       NApct_lim=20,
                       rmNApct=FALSE)
# Missing values represented by NA are handled.
print(dataEX_tmp)
#> # A tibble: 246 × 4
#>    id      time       X_month_state2 NApct
#>    <chr>   <date>              <dbl> <dbl>
#>  1 serie 1 2000-02-01        -0.0883   6.6
#>  2 serie 1 2000-03-01         0.205    6.6
#>  3 serie 1 2000-04-01         0.837    6.6
#>  4 serie 1 2000-05-01         0.737    6.6
#>  5 serie 1 2000-06-01         1.35    19.7
#>  6 serie 1 2000-07-01         1.55     9.9
#>  7 serie 1 2000-08-01         2.39    16.4
#>  8 serie 1 2000-09-01         2.41    13.1
#>  9 serie 1 2000-10-01         2.56     9.9
#> 10 serie 1 2000-11-01        NA       23  
#> # ℹ 236 more rows
# In this second step, for the yearly maximum, the sampling period
# for each year is modified to a fixed value.
dataEX_tmp =
    process_extraction(data=dataEX_tmp,
                       funct=list(XX_state2=max),
                       funct_args=list("X_month_state2", na.rm=TRUE),
                       sampling_period=c("05-01", "11-30"),
                       time_step="year")

# Finaly, the median is computed
process_extraction(data=dataEX_tmp,
                   funct=list(med_XX_state2=median),
                   funct_args=list("XX_state2", na.rm=TRUE),
                   time_step="none")
#> # A tibble: 2 × 2
#>   id      med_XX_state2
#>   <chr>           <dbl>
#> 1 serie 1          19.3
#> 2 serie 2          18.7

## Extraction of the monthly average and the monthly maximum in a
# single call.
# The output is in long tibble format with 12 values for each time
# serie.
process_extraction(data=data,
                   funct=list(XA_state1=mean,
                              XX_state2=max),
                   funct_args=list(list("X_state1", na.rm=TRUE),
                                   list("X_state2", na.rm=TRUE)),
                   time_step="month")
#> # A tibble: 24 × 5
#>    id      time       Month XA_state1 XX_state2
#>    <chr>   <date>     <dbl>     <dbl>     <dbl>
#>  1 serie 1 2000-01-01     1    22997.      38.3
#>  2 serie 1 2000-02-01     2    22933.      39.0
#>  3 serie 1 2000-03-01     3    22815.      40.0
#>  4 serie 1 2000-04-01     4    22774.      37.7
#>  5 serie 1 2000-05-01     5    22716.      35.4
#>  6 serie 1 2000-06-01     6    22747.      36.6
#>  7 serie 1 2000-07-01     7    22805.      36.5
#>  8 serie 1 2000-08-01     8    22755.      36.9
#>  9 serie 1 2000-09-01     9    22867.      37.1
#> 10 serie 1 2000-10-01    10    22921.      38.3
#> # ℹ 14 more rows
# In this case, the output tibble is compress to have the date
# indication in column. This also works for year-month extraction.
process_extraction(data=data,
                   funct=list(XA_state1=mean,
                              XX_state2=max),
                   funct_args=list(list("X_state1", na.rm=TRUE),
                                   list("X_state2", na.rm=TRUE)),
                   time_step="month",
                   compress=TRUE)
#> # A tibble: 2 × 25
#>   id      XA_state1_jan XA_state1_feb XA_state1_mar XA_state1_apr XA_state1_may
#>   <chr>           <dbl>         <dbl>         <dbl>         <dbl>         <dbl>
#> 1 serie 1        22997.        22933.        22815.        22774.        22716.
#> 2 serie 2        23007.        22869.        22925.        22724.        22785.
#> # ℹ 19 more variables: XA_state1_jun <dbl>, XA_state1_jul <dbl>,
#> #   XA_state1_aug <dbl>, XA_state1_sep <dbl>, XA_state1_oct <dbl>,
#> #   XA_state1_nov <dbl>, XA_state1_dec <dbl>, XX_state2_jan <dbl>,
#> #   XX_state2_feb <dbl>, XX_state2_mar <dbl>, XX_state2_apr <dbl>,
#> #   XX_state2_may <dbl>, XX_state2_jun <dbl>, XX_state2_jul <dbl>,
#> #   XX_state2_aug <dbl>, XX_state2_sep <dbl>, XX_state2_oct <dbl>,
#> #   XX_state2_nov <dbl>, XX_state2_dec <dbl>
# And in this last case, the output tibble is compress and then
# expand in order to have a list of tibble for each variable.
process_extraction(data=data,
                   funct=list(XA_state1=mean,
                              XX_state2=max),
                   funct_args=list(list("X_state1", na.rm=TRUE),
                                   list("X_state2", na.rm=TRUE)),
                   time_step="month",
                   compress=TRUE,
                   expand=TRUE)
#> $XA_state1_jan
#> # A tibble: 2 × 2
#>   id      XA_state1_jan
#>   <chr>           <dbl>
#> 1 serie 1        22997.
#> 2 serie 2        23007.
#> 
#> $XA_state1_feb
#> # A tibble: 2 × 2
#>   id      XA_state1_feb
#>   <chr>           <dbl>
#> 1 serie 1        22933.
#> 2 serie 2        22869.
#> 
#> $XA_state1_mar
#> # A tibble: 2 × 2
#>   id      XA_state1_mar
#>   <chr>           <dbl>
#> 1 serie 1        22815.
#> 2 serie 2        22925.
#> 
#> $XA_state1_apr
#> # A tibble: 2 × 2
#>   id      XA_state1_apr
#>   <chr>           <dbl>
#> 1 serie 1        22774.
#> 2 serie 2        22724.
#> 
#> $XA_state1_may
#> # A tibble: 2 × 2
#>   id      XA_state1_may
#>   <chr>           <dbl>
#> 1 serie 1        22716.
#> 2 serie 2        22785.
#> 
#> $XA_state1_jun
#> # A tibble: 2 × 2
#>   id      XA_state1_jun
#>   <chr>           <dbl>
#> 1 serie 1        22747.
#> 2 serie 2        22771.
#> 
#> $XA_state1_jul
#> # A tibble: 2 × 2
#>   id      XA_state1_jul
#>   <chr>           <dbl>
#> 1 serie 1        22805.
#> 2 serie 2        22832.
#> 
#> $XA_state1_aug
#> # A tibble: 2 × 2
#>   id      XA_state1_aug
#>   <chr>           <dbl>
#> 1 serie 1        22755.
#> 2 serie 2        22858.
#> 
#> $XA_state1_sep
#> # A tibble: 2 × 2
#>   id      XA_state1_sep
#>   <chr>           <dbl>
#> 1 serie 1        22867.
#> 2 serie 2        22805.
#> 
#> $XA_state1_oct
#> # A tibble: 2 × 2
#>   id      XA_state1_oct
#>   <chr>           <dbl>
#> 1 serie 1        22921.
#> 2 serie 2        22824.
#> 
#> $XA_state1_nov
#> # A tibble: 2 × 2
#>   id      XA_state1_nov
#>   <chr>           <dbl>
#> 1 serie 1        22930.
#> 2 serie 2        22854.
#> 
#> $XA_state1_dec
#> # A tibble: 2 × 2
#>   id      XA_state1_dec
#>   <chr>           <dbl>
#> 1 serie 1        22995.
#> 2 serie 2        22954.
#> 
#> $XX_state2_jan
#> # A tibble: 2 × 2
#>   id      XX_state2_jan
#>   <chr>           <dbl>
#> 1 serie 1          38.3
#> 2 serie 2          37.8
#> 
#> $XX_state2_feb
#> # A tibble: 2 × 2
#>   id      XX_state2_feb
#>   <chr>           <dbl>
#> 1 serie 1          39.0
#> 2 serie 2          38.3
#> 
#> $XX_state2_mar
#> # A tibble: 2 × 2
#>   id      XX_state2_mar
#>   <chr>           <dbl>
#> 1 serie 1          40.0
#> 2 serie 2          39.1
#> 
#> $XX_state2_apr
#> # A tibble: 2 × 2
#>   id      XX_state2_apr
#>   <chr>           <dbl>
#> 1 serie 1          37.7
#> 2 serie 2          37.1
#> 
#> $XX_state2_may
#> # A tibble: 2 × 2
#>   id      XX_state2_may
#>   <chr>           <dbl>
#> 1 serie 1          35.4
#> 2 serie 2          36.2
#> 
#> $XX_state2_jun
#> # A tibble: 2 × 2
#>   id      XX_state2_jun
#>   <chr>           <dbl>
#> 1 serie 1          36.6
#> 2 serie 2          35.3
#> 
#> $XX_state2_jul
#> # A tibble: 2 × 2
#>   id      XX_state2_jul
#>   <chr>           <dbl>
#> 1 serie 1          36.5
#> 2 serie 2          36.7
#> 
#> $XX_state2_aug
#> # A tibble: 2 × 2
#>   id      XX_state2_aug
#>   <chr>           <dbl>
#> 1 serie 1          36.9
#> 2 serie 2          36.9
#> 
#> $XX_state2_sep
#> # A tibble: 2 × 2
#>   id      XX_state2_sep
#>   <chr>           <dbl>
#> 1 serie 1          37.1
#> 2 serie 2          37.4
#> 
#> $XX_state2_oct
#> # A tibble: 2 × 2
#>   id      XX_state2_oct
#>   <chr>           <dbl>
#> 1 serie 1          38.3
#> 2 serie 2          36.8
#> 
#> $XX_state2_nov
#> # A tibble: 2 × 2
#>   id      XX_state2_nov
#>   <chr>           <dbl>
#> 1 serie 1          37.1
#> 2 serie 2          37.7
#> 
#> $XX_state2_dec
#> # A tibble: 2 × 2
#>   id      XX_state2_dec
#>   <chr>           <dbl>
#> 1 serie 1          38.3
#> 2 serie 2          37.6
#> 

## Extraction of the seasonal average and season maxium for both
# columns using suffixes to avoid repetition with compress and
# expand formating.
process_extraction(data=data,
                   funct=list(XA=mean,
                              XX=max),
                   funct_args=list(list("X", na.rm=TRUE),
                                   list("X", na.rm=TRUE)),
                   suffix=c("state1", "state2"),
                   time_step="season",
                   compress=TRUE,
                   expand=TRUE)
#> $XA_MAM
#> # A tibble: 2 × 3
#>   id      XA_MAM_state1 XA_MAM_state2
#>   <chr>           <dbl>         <dbl>
#> 1 serie 1        22770.          17.8
#> 2 serie 2        22816.          18.1
#> 
#> $XA_JJA
#> # A tibble: 2 × 3
#>   id      XA_JJA_state1 XA_JJA_state2
#>   <chr>           <dbl>         <dbl>
#> 1 serie 1        22769.          18.0
#> 2 serie 2        22821.          18.1
#> 
#> $XA_SON
#> # A tibble: 2 × 3
#>   id      XA_SON_state1 XA_SON_state2
#>   <chr>           <dbl>         <dbl>
#> 1 serie 1        22906.          18.9
#> 2 serie 2        22828.          19.0
#> 
#> $XA_DJF
#> # A tibble: 2 × 3
#>   id      XA_DJF_state1 XA_DJF_state2
#>   <chr>           <dbl>         <dbl>
#> 1 serie 1        22975.          19.4
#> 2 serie 2        22943.          19.3
#> 
#> $XX_MAM
#> # A tibble: 2 × 3
#>   id      XX_MAM_state1 XX_MAM_state2
#>   <chr>           <dbl>         <dbl>
#> 1 serie 1        27793.          40.0
#> 2 serie 2        26743.          39.1
#> 
#> $XX_JJA
#> # A tibble: 2 × 3
#>   id      XX_JJA_state1 XX_JJA_state2
#>   <chr>           <dbl>         <dbl>
#> 1 serie 1        27388.          36.9
#> 2 serie 2        27411.          36.9
#> 
#> $XX_SON
#> # A tibble: 2 × 3
#>   id      XX_SON_state1 XX_SON_state2
#>   <chr>           <dbl>         <dbl>
#> 1 serie 1        26734.          38.3
#> 2 serie 2        26829.          37.7
#> 
#> $XX_DJF
#> # A tibble: 2 × 3
#>   id      XX_DJF_state1 XX_DJF_state2
#>   <chr>           <dbl>         <dbl>
#> 1 serie 1        27316.          39.0
#> 2 serie 2        28217.          38.3
#>