process_trend

Process a trend analyze on time series data. The Mann-Kendall statistical test is applied to detect trends and some additional statistics are computed like the Sen-Theil slope estimator.

Usage

process_trend(
  dataEX,
  MK_level = 0.1,
  time_dependency_option = "INDE",
  suffix = NULL,
  suffix_delimiter = "_",
  to_normalise = FALSE,
  metaEX = NULL,
  extreme_take_not_signif_into_account = TRUE,
  extreme_take_only_series = NULL,
  extreme_by_suffix = TRUE,
  period_trend = NULL,
  period_change = NULL,
  extreme_prob = 0.01,
  show_advance_stat = FALSE,
  dev = FALSE,
  verbose = FALSE,
  verbose_stat = FALSE
)

Arguments

dataEX

Input data format is a tibble from the tibble package. It needs to have :

Only one column of Date that are regularly spaced and unique for each time serie.

If there is more than one time serie, at least one column needs to be of character string for names of time series in order to identify them. If more than one column of identifier is given, they will all be used in order to identify a unique time serie.

At least one column of numeric (or logical) on which the process of variable extraction will be perform. More numerical column can be leave but if they are useless, they will be suppressed. The tibble output of the process_extraction() function is the kind of wanted input data.

MK_level

numeric Mann-Kendall test significance level between 0 and 1. Default is 0.1.

time_dependency_option

character string for handling temporal dependence for the Mann-Kendall test. Possible values are :

"INDE", assume independence (i.e. the standard MK test)

"AR1", assumes AR1 short-term dependence structure (i.e. Hamed and Rao's version of the MK test)

"LTP", assume long-term persistence (i.e. Hamed's version of the MK test)

suffix

A character string vector representing suffixes to be appended to the column names of the extracted variables. See process_extraction() for more info. Default NULL.

suffix_delimiter

character string specifies the delimiter to use between the variable name and the suffix if not NULL. The default is "_".

to_normalise

A named logical vector indicating whether each variable's trend should be normalised. TRUE performs normalisation, while FALSE does nothing. This vector must be of length one or have the same length as the number of numeric (or logical) variables in dataEX, with names specifying which value corresponds to which variable. Default 'FALSE'.

metaEX

One of the outputs of the CARD_extraction function that contains metadata for the normalisation process. Default is NULL. If supplied, this normalisation information will be used instead of the settings provided in the to_normalise variable.

extreme_take_not_signif_into_account

logical Whether to consider non-significant trends in the computation of extreme trends. Default is TRUE.

extreme_take_only_series

character string A vector of the names of time series to be used for computing extreme trends. Default is NULL, which includes all available series.

extreme_by_suffix

logical If TRUE, extreme trends will be computed across separate sets of trend values of the same variable and the same suffix. If FALSE, all extreme trends of a variable will be used without considering suffixes. Default is TRUE.

period_trend

A vector of two dates (or two unambiguous character strings that can be coerced to dates) to restrict the period of analysis. As an example, it can be c("1950-01-01", "2020-12-31") to select data from the 1st January of 1950 to the end of December of 2020. The default option is period_trend=NULL, which considers all available data for each time serie.

period_change

in developpement A list of two vectors of two dates (or two unambiguous character strings that can be coerced to dates) to restrict two period for the change analysis. As an example, it can be list(c("1950-01-01", "1999-12-31"), c("2000-01-01", "2020-12-31")). The default option is period_change=NULL, which does not do any analysis.

extreme_prob

numeric The probability for identifying extreme trends using quantiles. Default is 0.01, so the computed extremes will be based on the quantile at 0.01 and 0.99.

show_advance_stat

logical Whether to display advanced statistical details. Default is FALSE.

dev

logical If TRUE, development mode is enabled. Default is FALSE.

verbose

logical Whether to print intermediate messages. Default is FALSE.

verbose_stat

logical Whether to print detailed statistical messages. Default is FALSE.

Value

A tibble with trend analysis results, including trend coefficients and statistical significance for each variables.

More precisely, the returned trendEX tibble contains the following columns :

* : The idenfiant of time series

variable_en : The name in english of variables

level : see MK_level

H : The result of the Mann-Kendall trend test. If TRUE a trend is detected at a type I error of MK_level (and a trust of 1-MK_level). If FALSE, a trend is NOT detected at a type I error of MK_level.

p: The p-value indicating the statistical significance of the Mann-Kendall trend test.

a : The Then-Seil slope estimator that gives an approximation of the trend coefficient. WARNING : A value is always return even if the Mann-Kendall trend test is not significant.

b : The ordinate at the origin in sort that you can trace `Y = a * X + b

period_trend : An information about the period of analyse for the trend

mean_period_trend : The average value of the variable along the period_trend that is usefull for normalisation (it is NA if to_normalise is FALSE for this variable).

a_normalise : The a value normalised with mean_period_trend.

a_normalise_min : The minimum of a_normalise values according to the selection made with extreme_take_not_signif_into_account, extreme_take_only_series and extreme_by_suffix.

a_normalise_max : Same as a_normalise_min but for maximum values.

If the the suffix argument is used, a column is added :

variable_no_suffix_en : The name in english of variables without suffixes

And also, if the period_change argument is filled up, more columns are added :

period_change : An information about the period of analyse for the change break

mean_period_change :

change :

change_min :

change_max :

See also

Examples

## Creation of random data set
set.seed(99)
Start = as.Date("2000-02-01")
End = as.Date("2010-04-04")
Date = seq.Date(Start, End, by="day")

# First time serie
data_1 = dplyr::tibble(time=Date,
                       X_state1=as.numeric(Date) +
                           rnorm(length(Date), 1e4, 1e3),
                       X_state2=seq(1, length(Date))/1e2 +
                           rnorm(length(Date), 0, 1),
                       id="serie 1")
data_1$X_state2[round(runif(500, 1, nrow(data_1)))] = NA

# Second time serie
data_2 = dplyr::tibble(time=Date,
                       X_state1=as.numeric(Date) +
                           rnorm(length(Date), 1e4, 1e3),
                       X_state2=seq(1, length(Date))/1e2 +
                           rnorm(length(Date), 0, 1),
                       id="serie 2")
data_2$X_state2[round(runif(1000, 1, nrow(data_2)))] = NA

# Final data for testing
data = dplyr::bind_rows(data_1, data_2)

## Extraction of the yearly average of daily value.
dataEX = process_extraction(data=data,
                   funct=list(XA_state1=mean),
                   funct_args=list("X_state1", na.rm=TRUE),
                   time_step="year")

dataEX = process_extraction(data=data,
                            funct=list(XA=mean,
                                       XX=max),
                            funct_args=list(list("X", na.rm=TRUE),
                                            list("X", na.rm=TRUE)),
                            suffix=c("state1", "state2"),
                            time_step="year")

## Trend test
# The direct application does not take care of possible grouped
# variables by suffix for extremes trends values.
trendEX1 = process_trend(dataEX,
                         to_normalise=TRUE)
#> Warning: 'to_normalise' is a unique value so it will be repeated for all the variables : XA_state1, XX_state1, XA_state2, XX_state2
print(trendEX1, width=Inf)
#> # A tibble: 8 × 12
#>   id      variable_en level H             p      a       b period_trend
#>   <chr>   <chr>       <dbl> <lgl>     <dbl>  <dbl>   <dbl> <list>      
#> 1 serie 1 XA_state1     0.1 TRUE  0.0000262 366.   10168.  <date [2]>  
#> 2 serie 1 XA_state2     0.1 TRUE  0.0000262   3.64  -108.  <date [2]>  
#> 3 serie 1 XX_state1     0.1 TRUE  0.000186  391.   12367.  <date [2]>  
#> 4 serie 1 XX_state2     0.1 TRUE  0.0000262   3.52   -99.8 <date [2]>  
#> 5 serie 2 XA_state1     0.1 TRUE  0.0000262 366.   10155.  <date [2]>  
#> 6 serie 2 XA_state2     0.1 TRUE  0.0000262   3.63  -107.  <date [2]>  
#> 7 serie 2 XX_state1     0.1 TRUE  0.000186  404.   11774.  <date [2]>  
#> 8 serie 2 XX_state2     0.1 TRUE  0.0000262   3.57  -102.  <date [2]>  
#>   mean_period_trend a_normalise a_normalise_min a_normalise_max
#>               <dbl>       <dbl>           <dbl>           <dbl>
#> 1           22959.         1.59            1.59            1.59
#> 2              19.7       18.5            18.5            18.5 
#> 3           26063.         1.50            1.50            1.56
#> 4              23.4       15.0            15.0            15.4 
#> 5           22967.         1.59            1.59            1.59
#> 6              19.6       18.5            18.5            18.5 
#> 7           25897.         1.56            1.50            1.56
#> 8              23.1       15.4            15.0            15.4 

# More complicated case with the use of 'period_change' and customs
# normalisation info
trendEX2 =
    process_trend(dataEX,
                  suffix=c("state1", "state2"),
                  suffix_delimiter="_",
                  to_normalise=c("XA_state1"=FALSE,
                                 "XA_state2"=FALSE,
                                 "XX_state1"=TRUE,
                                 "XX_state2"=TRUE),
                  extreme_take_only_series=NULL,
                  extreme_by_suffix=FALSE,
                  period_change=list(c(as.Date("2000-01-01"),
                                       as.Date("2005-01-01")),
                                     c(as.Date("2006-01-01"),
                                       as.Date("2010-01-01"))))
print(trendEX2, width=Inf)
#> # A tibble: 8 × 18
#>   id      variable_en variable_no_suffix_en level H             p      a       b
#>   <chr>   <chr>       <chr>                 <dbl> <lgl>     <dbl>  <dbl>   <dbl>
#> 1 serie 1 XA_state1   XA                      0.1 TRUE  0.0000262 366.   10168. 
#> 2 serie 1 XA_state2   XA                      0.1 TRUE  0.0000262   3.64  -108. 
#> 3 serie 1 XX_state1   XX                      0.1 TRUE  0.000186  391.   12367. 
#> 4 serie 1 XX_state2   XX                      0.1 TRUE  0.0000262   3.52   -99.8
#> 5 serie 2 XA_state1   XA                      0.1 TRUE  0.0000262 366.   10155. 
#> 6 serie 2 XA_state2   XA                      0.1 TRUE  0.0000262   3.63  -107. 
#> 7 serie 2 XX_state1   XX                      0.1 TRUE  0.000186  404.   11774. 
#> 8 serie 2 XX_state2   XX                      0.1 TRUE  0.0000262   3.57  -102. 
#>   period_trend mean_period_trend a_normalise a_normalise_min a_normalise_max
#>   <list>                   <dbl>       <dbl>           <dbl>           <dbl>
#> 1 <date [2]>                NA        366.              3.63           366. 
#> 2 <date [2]>                NA          3.64            3.63           366. 
#> 3 <date [2]>             26063.         1.50            1.50            15.4
#> 4 <date [2]>                23.4       15.0             1.50            15.4
#> 5 <date [2]>                NA        366.              3.63           366. 
#> 6 <date [2]>                NA          3.63            3.63           366. 
#> 7 <date [2]>             25897.         1.56            1.50            15.4
#> 8 <date [2]>                23.1       15.4             1.50            15.4
#>   period_change mean_period_change  change change_min change_max
#>   <list>        <list>               <dbl>      <dbl>      <dbl>
#> 1 <list [2]>    <dbl [2]>          1985.        19.8       2025.
#> 2 <list [2]>    <dbl [2]>            19.8       19.8       2025.
#> 3 <list [2]>    <dbl [2]>             8.33       7.68       132.
#> 4 <list [2]>    <dbl [2]>           132.         7.68       132.
#> 5 <list [2]>    <dbl [2]>          2026.        19.8       2025.
#> 6 <list [2]>    <dbl [2]>            19.8       19.8       2025.
#> 7 <list [2]>    <dbl [2]>             7.66       7.68       132.
#> 8 <list [2]>    <dbl [2]>           132.         7.68       132.